feat: move SwapKitWidget to @swapkit/ui/react
