+++ playgrounds/nextjs/src/app/page.tsx
@@ -1,191 +1,9 @@
 "use client";
 
-import { AssetValue, type Chain } from "@swapkit/helpers";
-import type { QuoteResponseRoute } from "@swapkit/helpers/api";
-import { ProviderName, SwapKitApi } from "@swapkit/sdk";
-import { ArrowDownUp, Loader2 } from "lucide-react";
-import { useCallback, useEffect, useState } from "react";
-import { toast } from "sonner";
-import { Button } from "~/components/ui/button";
-import { Card, CardContent } from "~/components/ui/card";
-import { useSwapKit } from "~/lib/swapKit";
+import { SwapKitWidget } from "@swapkit/ui/react";
+
+import "@swapkit/ui/swapkit.css";
 
 export default function SwapPage() {
-  const { swapKit, isWalletConnected } = useSwapKit();
-  const [inputAsset, setInputAsset] = useState<string>("NEAR.USDT-usdt.tether-token.near");
-  const [outputAsset, setOutputAsset] = useState<string>("THOR.RUNE");
-  const [amount, setAmount] = useState("4.20");
-  const [isSwapping, setIsSwapping] = useState(false);
-  const [estimatedOutput, setEstimatedOutput] = useState<string>();
-  const [routes, setRoutes] = useState<QuoteResponseRoute[]>([]);
-
-  const handleSwap = async (route: QuoteResponseRoute) => {
-    if (!swapKit) return;
-
-    try {
-      setIsSwapping(true);
-      const swap = await swapKit.swap({ route });
-
-      await swap.wait();
-      setAmount("");
-      setEstimatedOutput(undefined);
-      setRoutes([]);
-      toast.success("Swap completed successfully");
-    } catch (error) {
-      console.error("Swap failed:", error);
-      toast.error(`Swap failed: ${error instanceof Error ? error.message : "Unknown error"}`);
-    } finally {
-      setIsSwapping(false);
-    }
-  };
-
-  const swap = async (route: QuoteResponseRoute, inputAssetValue?: AssetValue) => {
-    if (!(inputAssetValue && swapKit)) return;
-
-    try {
-      const isChainflip = route?.providers?.includes(ProviderName.CHAINFLIP);
-      if (isChainflip) {
-        await handleSwap(route);
-        return;
-      }
-
-      const tx = route.tx;
-      if (!tx || typeof tx === "string" || !("from" in tx)) {
-        throw new Error("Invalid transaction format");
-      }
-
-      const isApproved = await swapKit.isAssetValueApproved(inputAssetValue, tx.from);
-      if (isApproved) {
-        await handleSwap(route);
-      } else {
-        await swapKit.approveAssetValue(inputAssetValue, tx.from);
-        toast.success("Asset approved, you can now swap");
-      }
-    } catch (error) {
-      console.error("Swap process failed:", error);
-      toast.error(`Swap process failed: ${error instanceof Error ? error.message : "Unknown error"}`);
-    }
-  };
-
-  const updateEstimatedOutput = useCallback(async () => {
-    if (!(inputAsset && outputAsset && amount && swapKit)) {
-      setEstimatedOutput(undefined);
-      setRoutes([]);
-      return;
-    }
-
-    try {
-      const sourceAddress = swapKit.getAddress(inputAsset.split(".")[0] as Chain);
-      const destinationAddress = swapKit.getAddress(outputAsset.split(".")[0] as Chain);
-
-      const quote = await SwapKitApi.getSwapQuote({
-        buyAsset: outputAsset,
-        destinationAddress,
-        includeTx: true,
-        sellAmount: amount,
-        sellAsset: inputAsset,
-        slippage: 3,
-        sourceAddress,
-      });
-
-      if (quote?.routes?.length) {
-        setRoutes(quote.routes);
-        setEstimatedOutput(quote.routes[0].expectedBuyAmount);
-      }
-    } catch (error) {
-      console.error("Failed to get quote:", error);
-      toast.error(`Failed to get quote: ${error instanceof Error ? error.message : "Unknown error"}`);
-      setEstimatedOutput(undefined);
-      setRoutes([]);
-    }
-  }, [inputAsset, outputAsset, amount, swapKit]);
-
-  useEffect(() => {
-    updateEstimatedOutput();
-  }, [updateEstimatedOutput]);
-
-  return (
-    <div className="flex flex-col gap-4">
-      <h1 className="font-medium text-2xl">Swap</h1>
-
-      <Card>
-        <CardContent className="grid gap-6">
-          <div className="space-y-4">
-            <div className="grid gap-4">
-              <SwapInputWithChainSelector
-                amount={amount}
-                isSwapping={isSwapping}
-                label="Pay"
-                selectedAsset={inputAsset}
-                setAmount={setAmount}
-                setSelectedAsset={setInputAsset}
-              />
-
-              <div className="-my-4 flex items-center space-x-4">
-                <span className="h-px w-full bg-border" />
-
-                <Button
-                  className="size-10 shrink-0 rounded-full"
-                  onClick={() => {
-                    const temp = inputAsset;
-                    setInputAsset(outputAsset);
-                    setOutputAsset(temp);
-                  }}
-                  size="unstyled"
-                  variant="tertiary">
-                  <ArrowDownUp className="size-6" />
-                </Button>
-
-                <span className="h-px w-full bg-border" />
-              </div>
-
-              <SwapInputWithChainSelector
-                amount={estimatedOutput}
-                isSwapping={isSwapping}
-                label="Receive"
-                selectedAsset={outputAsset}
-                setSelectedAsset={setOutputAsset}
-              />
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-
-      <Button
-        className="w-full"
-        disabled={!(inputAsset && outputAsset && amount) || isSwapping || !isWalletConnected}
-        onClick={async () => {
-          if (!(routes.length && inputAsset)) return;
-          try {
-            const assetValue = await AssetValue.from({ amount, asset: inputAsset, asyncTokenLookup: true });
-            const amountValue = assetValue.set(amount);
-            await swap(routes[0], amountValue);
-          } catch (error) {
-            console.error("Failed to prepare swap:", error);
-            toast.error(`Failed to prepare swap: ${error instanceof Error ? error.message : "Unknown error"}`);
-          }
-        }}
-        size="xl"
-        variant="primary">
-        {isSwapping ? (
-          <>
-            <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-            Swapping...
-          </>
-        ) : isWalletConnected ? (
-          inputAsset && outputAsset ? (
-            amount ? (
-              "Swap"
-            ) : (
-              "Enter Amount"
-            )
-          ) : (
-            "Select Assets"
-          )
-        ) : (
-          "Connect wallet"
-        )}
-      </Button>
-    </div>
-  );
+  return <SwapKitWidget apiKey={process.env.NEXT_PUBLIC_TEST_API_KEY || ""} />;
 }
